refactor: constants/api를 역할에 따라 분리
