+++ src/constants/api/errorCode.ts
@@ -1,0 +1,34 @@
+export const VALIDATION_ERROR_CODE = 30001;
+
+export const AUTH_ERROR_CODE = {
+	INVALID_VERIFY_TOKEN: 40101,
+	INVALID_EMAIL_OR_PASSWORD: 40102,
+	UNAUTHORIZED_OR_EXPIRED_VERIFY_TOKEN: 40103,
+	DUPLICATED_USER_EMAIL: 40104,
+	INVALID_OAUTH_PROVIDER: 40105,
+	INVALID_AUTHORIZATION_CODE: 40106,
+	SOCIAL_EMAIL_ALREADY_REGISTERED: 40107,
+	MISSING_REQUIRED_USER_INFO: 40189,
+} as const;
+
+export const TOKEN_ERROR_CODE = {
+	FORBIDDEN_ACCESS_TOKEN: 40181,
+	EMPTY_ACCESS_TOKEN: 40182,
+	EXPIRED_ACCESS_TOKEN: 40183,
+	MALFORMED_TOKEN: 40184,
+	TAMPERED_TOKEN: 40185,
+	UNSUPPORTED_JWT_TOKEN: 40186,
+	TAKEN_AWAY_TOKEN: 40187,
+	EXPIRED_REFRESH_TOKEN: 40188,
+} as const;
+
+export const SESSION_INVALID_CODES = new Set<number>(
+	Object.values(TOKEN_ERROR_CODE).filter((code) => code !== TOKEN_ERROR_CODE.EXPIRED_ACCESS_TOKEN)
+);
+
+export const ABNORMAL_TOKEN_CODES = new Set<number>([
+	TOKEN_ERROR_CODE.MALFORMED_TOKEN,
+	TOKEN_ERROR_CODE.TAMPERED_TOKEN,
+	TOKEN_ERROR_CODE.UNSUPPORTED_JWT_TOKEN,
+	TOKEN_ERROR_CODE.TAKEN_AWAY_TOKEN,
+]);
